requirements.txt
Flask~=1.1.4
mysql-connector-python==8.0.19